==> part2/code/02. Verse Maker.ipynb <==
import pandas as pd
bible_db = pd.read_csv ('bible_db.csv')
bible_db


bible_db['Book'].unique()

old = ['Genesis', 'Exodus', 'Leviticus', 'Numbers', 'Deuteronomy',
       'Joshua', 'Judges', 'Ruth', '1 Samuel', '2 Samuel', '1 Kings',
       '2 Kings', '1 Chronicles', '2 Chronicles', 'Ezra', 'Nehemiah',
       'Esther', 'Job', 'Psalm', 'Proverbs', 'Ecclesiastes',
       'Song of Solomon', 'Isaiah', 'Jeremiah', 'Lamentations', 'Ezekiel',
       'Daniel', 'Hosea', 'Joel', 'Amos', 'Obadiah', 'Jonah', 'Micah',
       'Nahum', 'Habakkuk', 'Zephaniah', 'Haggai', 'Zechariah', 'Malachi']
new = ['Matthew', 'Mark', 'Luke', 'John', 'Acts', 'Romans',
       '1 Corinthians', '2 Corinthians', 'Galatians', 'Ephesians',
       'Philippians', 'Colossians', '1 Thessalonians', '2 Thessalonians',
       '1 Timothy', '2 Timothy', 'Titus', 'Philemon', 'Hebrews', 'James',
       '1 Peter', '2 Peter', '1 John', '2 John', '3 John', 'Jude',
       'Revelation', 'Tobit', 'Judith', 'Greek Esther', 'Wisdom',
       'Sirach', 'Baruch', 'Letter of Jeremiah', 'Prayer of Azariah',
       'Susanna', 'Bel and the Dragon', '1 Maccabees', '2 Maccabees',
       '1 Esdras', 'Prayer of Manasseh', 'Psalm 151', '3 Maccabees',
       '2 Esdras', '4 Maccabees']

new_abridged = ['Matthew', 'Mark', 'Luke', 'John', 'Acts', 'Romans',
       '1 Corinthians', '2 Corinthians', 'Galatians', 'Ephesians',
       'Philippians', 'Colossians', '1 Thessalonians', '2 Thessalonians',
       '1 Timothy', '2 Timothy', 'Titus', 'Philemon', 'Hebrews', 'James',
       '1 Peter', '2 Peter', '1 John', '2 John', '3 John', 'Jude',
       'Revelation']

keywords = ['love, hope', 'heart']

new_test = bible_db[bible_db['cleaned_verse'].str.contains('love')]
new_test

new_test = bible_db[bible_db['Book'] == 'Psalm']
new_test

new_test[new_test['cleaned_verse'].str.contains('I will extol you')]['cleaned_verse']

new_test[new_test['cleaned_verse'].str.contains('I will tell of the decree of the Lord ')]

#I will tell of the decree of the Lord are right the fear of the Lord

new_test[new_test['cleaned_verse'].str.contains('lion')]

bible_db[bible_db['cleaned_verse'].str.contains(' you will be astonished')]

import re
def text_cleaner(text):
    text = re.sub(r'--', ' ', text)
    text = re.sub(r';', '', text)
    text = re.sub(r',', '', text)
    text = re.sub(r'T', 't', text)
    text = re.sub('[\[].*?[\]]', '', text)
    #text = re.sub(r'(\b|\s+\-?|^\-?)(\d+|\d*\.\d+)\b','', text)
    #text = re.sub('[^a-zA-Z0-9\\s]+', '', text)
    text = ' '.join(text.split())
    return text

new_test_combine = [' '.join(new_test['cleaned_verse'])]
new_test_combine = new_test_combine[0]
new_test_verses = text_cleaner(new_test_combine)

import spacy
from spacy.lang.en import English

nlp = spacy.load('en_core_web_sm')
bible_doc = nlp(new_test_verses)
bible_sent = ' '.join([sent.text for sent in bible_doc.sents if len(sent.text) > 1])
bible_sent

import markovify
generator_1 = markovify.Text(bible_sent, state_size=3)

for i in range(3):
    print(generator_1.make_sentence())
#We will randomly generate three more sentences of no more than 100 characters
for i in range(3):
    print(generator_1.make_short_sentence(max_chars=100))

class POSifiedText(markovify.Text):
    def word_split(self, sentence):
        return ['::'.join((word.orth_, word.pos_)) for word in nlp(sentence)]
    def word_join(self, words):
        sentence = ' '.join(word.split('::')[0] for word in words)
        return sentence
#Call the class on our text
generator_2 = POSifiedText(bible_sent, state_size=3)

for i in range(5):
    print(generator_2.make_sentence())
#print 100 characters or less sentences

for i in range(5):
    
    print(generator_2.make_short_sentence(max_chars=100))

import nltk
from nltk.corpus import stopwords
nltk.download('stopwords')
from nltk.tokenize import word_tokenize



stop = stopwords.words('english')
bible_db["verse_stopped"] = bible_db["cleaned_verse"].apply(lambda x: ' '.join([word for word in x.split() if word.lower() not in (stop)]))

stop

bible_db["verse_stopped"].head()

from collections import Counter
Counter(" ".join(bible_db["verse_stopped"]).split()).most_common(100)
