==> part2/code/03. Prepping Training Testing Data .ipynb <==
import pandas as pd
bible_db = pd.read_csv ('bible_training_data.csv')
bible_db


bible_db['cleaned_verse'] = bible_db['cleaned_verse'].str.replace("\r", " ")
bible_db['cleaned_verse'] = bible_db['cleaned_verse'].str.replace("\n", " ")
bible_db['cleaned_verse'] = bible_db['cleaned_verse'].str.replace("    ", " ")
bible_db['cleaned_verse'] = bible_db['cleaned_verse'].str.replace("'s", "")
bible_db['cleaned_verse'] = bible_db['cleaned_verse'].str.lower()
punctuation_signs = list("?:!.,;")

for punct_sign in punctuation_signs:
    bible_db['cleaned_verse'] = bible_db['cleaned_verse'].str.replace(punct_sign, '')
    
bible_db


from nltk.stem import WordNetLemmatizer

# Saving the lemmatizer into an object
wordnet_lemmatizer = WordNetLemmatizer()
nrows = len(bible_db)
lemmatized_text_list = []

for row in range(0, nrows):
    
    # Create an empty list containing lemmatized words
    lemmatized_list = []
    
    # Save the text and its words into an object
    text = bible_db.loc[row]['cleaned_verse']
    text_words = text.split(" ")

    # Iterate through every word to lemmatize
    for word in text_words:
        lemmatized_list.append(wordnet_lemmatizer.lemmatize(word, pos="v"))
        
    # Join the list
    lemmatized_text = " ".join(lemmatized_list)
    
    # Append to the list containing the texts
    lemmatized_text_list.append(lemmatized_text)
bible_db['lemma_verses'] = lemmatized_text_list
bible_db


from nltk.corpus import stopwords
stop_words = list(stopwords.words('english'))

for stop_word in stop_words:

    regex_stopword = r"\b" + stop_word + r"\b"
    bible_db['lemma_verses'] = bible_db['lemma_verses'].str.replace(regex_stopword, '')
bible_db

bible_db.Category.unique()

category_codes = {
    'Law': 0,
    'Historical': 1,
    'Prophets': 2,
    'Gospel': 3,
    'Letters': 4
}

bible_db['Category_Code'] = bible_db['Category']
bible_db = bible_db.replace({'Category_Code':category_codes})
bible_db

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(bible_db['lemma_verses'], 
                                                    bible_db['Category_Code'], 
                                                    test_size=0.15, 
                                                    random_state=8)

# Parameter election
ngram_range = (1,2)
min_df = 10
max_df = 1.
max_features = 300

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_selection import chi2
tfidf = TfidfVectorizer(encoding='utf-8',
                        ngram_range=ngram_range,
                        stop_words=None,
                        lowercase=False,
                        max_df=max_df,
                        min_df=min_df,
                        max_features=max_features,
                        norm='l2',
                        sublinear_tf=True)
                        
features_train = tfidf.fit_transform(X_train).toarray()
labels_train = y_train
print(features_train.shape)

features_test = tfidf.transform(X_test).toarray()
labels_test = y_test
print(features_test.shape)

from sklearn.feature_selection import chi2
import numpy as np

for Product, category_id in sorted(category_codes.items()):
    features_chi2 = chi2(features_train, labels_train == category_id)
    indices = np.argsort(features_chi2[0])
    feature_names = np.array(tfidf.get_feature_names())[indices]
    unigrams = [v for v in feature_names if len(v.split(' ')) == 1]
    bigrams = [v for v in feature_names if len(v.split(' ')) == 2]
    print("# '{}' category:".format(Product))
    print("  . Most correlated unigrams:\n. {}".format('\n. '.join(unigrams[-5:])))
    print("  . Most correlated bigrams:\n. {}".format('\n. '.join(bigrams[-2:])))
    print("")

import pickle
# X_train
with open('Pickles/X_train.pickle', 'wb') as output:
    pickle.dump(X_train, output)
    
# X_test    
with open('Pickles/X_test.pickle', 'wb') as output:
    pickle.dump(X_test, output)
    
# y_train
with open('Pickles/y_train.pickle', 'wb') as output:
    pickle.dump(y_train, output)
    
# y_test
with open('Pickles/y_test.pickle', 'wb') as output:
    pickle.dump(y_test, output)
    
# df
with open('Pickles/df.pickle', 'wb') as output:
    pickle.dump(bible_db, output)
    
# features_train
with open('Pickles/features_train.pickle', 'wb') as output:
    pickle.dump(features_train, output)

# labels_train
with open('Pickles/labels_train.pickle', 'wb') as output:
    pickle.dump(labels_train, output)

# features_test
with open('Pickles/features_test.pickle', 'wb') as output:
    pickle.dump(features_test, output)

# labels_test
with open('Pickles/labels_test.pickle', 'wb') as output:
    pickle.dump(labels_test, output)
    
# TF-IDF object
with open('Pickles/tfidf.pickle', 'wb') as output:
    pickle.dump(tfidf, output)
